리뷰 반영 및 2-5 구현
리뷰 반영
1. count(*) -> count(id) 수정 article, reply
2. Article이 Reply를 리스트로 가지도록 수정
3. 삭제하지 않았던 test-data.sql 삭제
4. 잔실수 수정

2-5 구현
1. 페이징 구현
--- 5개 단위로 끊어서 표현 (ex. 3페이지를 조회하면 1~5 페이지로 이동할수 있는 목록이, 7페이지를 조회하면 6~10 페이지로 이동할수 있는 목록이 표시됨)
--- 앞 뒤로 이동할 수 있는 버튼 추가 (ex. 3페이지 에서 앞 버튼 누르면 4페이지 이동 6페이지에서 뒤 버튼 누르면 5페이지로 이동, 첫페이지와 마지막 페이지는 각각 뒤 버튼, 앞 버튼 X)

+++ src/main/java/com/kakao/cafe/service/article/ArticleService.java
@@ -3,12 +3,15 @@ package com.kakao.cafe.service.article;
 import com.kakao.cafe.dao.article.ArticleDao;
 import com.kakao.cafe.model.article.Article;
 import com.kakao.cafe.model.article.ArticleFactory;
+import com.kakao.cafe.service.Constant;
 import com.kakao.cafe.service.article.dto.ArticleCreateDto;
 import com.kakao.cafe.service.article.dto.ArticleDto;
 import com.kakao.cafe.service.article.dto.ArticleUpdateDto;
+import com.kakao.cafe.service.reply.dto.ReplyDto;
 import java.util.List;
 import java.util.stream.Collectors;
 import java.util.stream.IntStream;
+import javax.naming.NoPermissionException;
 import org.springframework.stereotype.Service;
 
 @Service
@@ -28,19 +31,36 @@ public class ArticleService {
                 .collect(Collectors.toList());
     }
 
-    public List<Integer> getPages(int articleLimit) {
+    public List<Integer> getPages(int articleLimit, int pageLimit, int page) {
+        int startIndex = calculatePageBaseNumber(page, pageLimit) + 1;
+        int finishIndex = calculatePageBaseNumber(page, pageLimit) + pageLimit;
+        int lastPageNumber = getLastPageNumber(articleLimit);
+
+        if (lastPageNumber < finishIndex) {
+            finishIndex = lastPageNumber;
+        }
+
         return IntStream
-                .rangeClosed(1, Math.floorDiv(articleDao.getSize() - 1, articleLimit) + 1)
+                .rangeClosed(startIndex, finishIndex)
                 .boxed()
                 .collect(Collectors.toList());
+    }
+
+    public int getLastPageNumber(int articleLimit) {
+        return (articleDao.getSize() - 1) / articleLimit + 1;
     }
 
     public void createArticle(ArticleCreateDto articleCreateDto) {
         articleDao.addArticle(ArticleFactory.getArticle(articleCreateDto));
     }
 
-    public void deleteArticle(int id) {
-        articleDao.deleteArticle(id);
+    public void deleteArticle(int id, String userId) throws NoPermissionException {
+        checkArticleWriter(id, userId);
+        if (isArticleHasOnlyUserIdReply(id, userId)) {
+            articleDao.deleteArticle(id);
+            return;
+        }
+        throw new NoPermissionException(Constant.OTHER_REPLY_EXIST);
     }
 
     public void updateArticle(int id, ArticleUpdateDto articleUpdateDto) {
@@ -55,15 +75,47 @@ public class ArticleService {
         return new ArticleDto(findArticle(id));
     }
 
+    public List<ReplyDto> getArticleReplies(int articleId) {
+        return articleDao
+                .findArticleById(articleId)
+                .orElseThrow(() -> new IllegalArgumentException(Constant.ARTICLE_IS_NOT_EXIST))
+                .getReplies()
+                .stream()
+                .map(ReplyDto::new)
+                .collect(Collectors.toList());
+    }
+
     private Article findArticle(int id) {
         return articleDao
                 .findArticleById(id)
-                .orElseThrow(() -> new IllegalArgumentException("찾는 게시물이 없습니다."));
+                .orElseThrow(() -> new IllegalArgumentException(Constant.ARTICLE_IS_NOT_EXIST));
     }
 
     private void checkExist(int id) {
         if (articleDao.findArticleById(id).isEmpty()) {
-            throw new IllegalArgumentException("찾는 게시물이 없습니다.");
+            throw new IllegalArgumentException(Constant.ARTICLE_IS_NOT_EXIST);
         }
+    }
+
+    private void checkArticleWriter(int id, String userId) throws NoPermissionException {
+        Article article = articleDao.findArticleById(id)
+                .orElseThrow(() -> new IllegalArgumentException(Constant.ARTICLE_IS_NOT_EXIST));
+
+        if (!article.getUserId().getValue().equals(userId)) {
+            throw new NoPermissionException(Constant.DELETE_ARTICLE_ONLY_WRITER);
+        }
+    }
+
+    private boolean isArticleHasOnlyUserIdReply(int articleId, String userId) {
+        return articleDao
+                .findArticleById(articleId)
+                .orElseThrow(() -> new IllegalArgumentException(Constant.ARTICLE_IS_NOT_EXIST))
+                .getReplies()
+                .stream()
+                .allMatch(reply -> reply.getUserId().getValue().equals(userId));
+    }
+
+    private int calculatePageBaseNumber(int page, int pageLimit) {
+        return Math.floorDiv(page - 1, pageLimit) * pageLimit;
     }
 }
